+++ include/parse_mrt.h
@@ -214,6 +214,6 @@ struct libparsebgp_parse_mrt_parsed_data {
  *
  * @return number of bytes read
  */
-int libparsebgp_parse_mrt_parse_msg(libparsebgp_parse_mrt_parsed_data *mrt_parsed_data, unsigned char *buffer, int buf_len);
+ssize_t libparsebgp_parse_mrt_parse_msg(libparsebgp_parse_mrt_parsed_data *mrt_parsed_data, unsigned char *buffer, int buf_len);
 
 #endif /* PARSEBMP_H_ */
